Split up source into more files.
